requirements.txt
openpyxl>=3.1.2
python-docx>=1.1.0
python-pptx>=0.6.23
anthropic>=0.40.0
yfinance>=0.2.40
requests>=2.31.0
matplotlib>=3.8.0
pandas>=2.2.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
Pillow>=10.0.0

# API service
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
# (Supabase Storage accessed via REST — no supabase-py dep needed)
